Extract Move functionality to Move Action script
