+++ Assets/Scripts/MoveAction.cs
@@ -1,0 +1,37 @@
+using UnityEngine;
+
+public class MoveAction : MonoBehaviour
+{
+    private Vector3 targetPosition;
+
+    public bool IsWalking { get; private set; }
+
+    private void Awake()
+    {
+        targetPosition = transform.position;
+    }
+
+    private void Update()
+    {
+        float stoppingDistance = 0.1f;
+        if (Vector3.Distance(targetPosition, transform.position) > stoppingDistance)
+        {
+            Vector3 moveDirection = (targetPosition - transform.position).normalized;
+            float moveSpeed = 4f;
+            transform.position += moveDirection * moveSpeed * Time.deltaTime;
+            IsWalking = true;
+
+            float rotateSpeed = 10f;
+            transform.forward = Vector3.Lerp(transform.forward, moveDirection, rotateSpeed * Time.deltaTime);
+        }
+        else
+        {
+            IsWalking = false;
+        }
+    }
+
+    public void Move(Vector3 targetPosition)
+    {
+        this.targetPosition = targetPosition;
+    }
+}
