+++ Assets/Scripts/Unit.cs
@@ -2,14 +2,12 @@ using UnityEngine;
 
 public class Unit : MonoBehaviour
 {
-    private Vector3 targetPosition;
+    private MoveAction moveAction;
     private GridPosition gridPosition;
-
-    public bool IsWalking { get; private set; }
 
     private void Awake()
     {
-        targetPosition = transform.position;
+        moveAction = GetComponent<MoveAction>();
     }
 
     private void Start()
@@ -20,21 +18,6 @@ public class Unit : MonoBehaviour
 
     private void Update()
     {
-        float stoppingDistance = 0.1f;
-        if(Vector3.Distance(targetPosition, transform.position) > stoppingDistance)
-        {
-            Vector3 moveDirection = (targetPosition - transform.position).normalized;
-            float moveSpeed = 4f;
-            transform.position += moveDirection * moveSpeed * Time.deltaTime;
-            IsWalking = true;
-
-            float rotateSpeed = 10f;
-            transform.forward = Vector3.Lerp(transform.forward, moveDirection, rotateSpeed * Time.deltaTime);
-        }
-        else
-        {
-            IsWalking=false;
-        }
 
         GridPosition newGridPosition = LevelGrid.Instance.GetGridPosition(transform.position);
         if(newGridPosition != gridPosition)
@@ -42,11 +25,10 @@ public class Unit : MonoBehaviour
             LevelGrid.Instance.MoveUnitToGridPosition(this, gridPosition, newGridPosition);
             gridPosition = newGridPosition;
         }
-
     }
 
-    public void Move(Vector3 targetPosition)
+    public MoveAction GetMoveAction()
     {
-        this.targetPosition = targetPosition;
+        return moveAction;
     }
 }
